Add test for cascade deletions
